+++ samples/Sentry.Samples.GraphQL.Server/Notes/NotesQuery.cs
@@ -1,0 +1,11 @@
+using GraphQL.Types;
+
+namespace Sentry.Samples.GraphQL.Server.Notes;
+
+public sealed class NotesQuery : ObjectGraphType
+{
+    public NotesQuery(NotesData data)
+    {
+        Field<ListGraphType<NoteType>>("notes").Resolve(context => data.GetAll());
+    }
+}