+++ samples/Sentry.Samples.GraphQL.Server/Telemetry.cs
@@ -1,0 +1,9 @@
+using System.Diagnostics;
+
+namespace Sentry.Samples.GraphQL.Server;
+
+public static class Telemetry
+{
+    public const string ServiceName = "Sentry.Samples.GraphQL";
+    public static ActivitySource ActivitySource { get; } = new(ServiceName);
+}