+++ src/Sentry/Protocol/Response.cs
@@ -43,6 +43,15 @@ public sealed class Response : IJsonSerializable, ICloneable<Response>, IUpdatab
     public string? Cookies { get; set; }
 
     /// <summary>
+    /// Submitted data in whatever format makes most sense.
+    /// </summary>
+    /// <remarks>
+    /// This data should not be provided by default as it can get quite large.
+    /// </remarks>
+    /// <value>The request payload.</value>
+    public object? Data { get; set; }
+
+    /// <summary>
     /// Gets or sets the headers.
     /// </summary>
     /// <remarks>
@@ -87,6 +96,7 @@ public sealed class Response : IJsonSerializable, ICloneable<Response>, IUpdatab
     {
         BodySize ??= source.BodySize;
         Cookies ??= source.Cookies;
+        Data ??= source.Data;
         StatusCode ??= source.StatusCode;
         source.InternalHeaders?.TryCopyTo(Headers);
     }
@@ -111,6 +121,7 @@ public sealed class Response : IJsonSerializable, ICloneable<Response>, IUpdatab
         writer.WriteString("type", Type);
         writer.WriteNumberIfNotNull("body_size", BodySize);
         writer.WriteStringIfNotWhiteSpace("cookies", Cookies);
+        writer.WriteDynamicIfNotNull("data", Data, logger);
         writer.WriteStringDictionaryIfNotEmpty("headers", InternalHeaders!);
         writer.WriteNumberIfNotNull("status_code", StatusCode);
 
@@ -124,6 +135,7 @@ public sealed class Response : IJsonSerializable, ICloneable<Response>, IUpdatab
     {
         var bodySize = json.GetPropertyOrNull("body_size")?.GetInt64();
         var cookies = json.GetPropertyOrNull("cookies")?.GetString();
+        var data = json.GetPropertyOrNull("data")?.GetDynamicOrNull();
         var headers = json.GetPropertyOrNull("headers")?.GetStringDictionaryOrNull();
         var statusCode = json.GetPropertyOrNull("status_code")?.GetInt16();
 
@@ -131,6 +143,7 @@ public sealed class Response : IJsonSerializable, ICloneable<Response>, IUpdatab
         {
             BodySize = bodySize,
             Cookies = cookies,
+            Data = data,
             InternalHeaders = headers?.WhereNotNullValue().ToDictionary(),
             StatusCode = statusCode
         };